Decompile DRA func_8011E4BC
